+++ src/test/java/wooteco/subway/domain/PathTest.java
@@ -5,6 +5,7 @@ import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.CsvSource;
 import wooteco.subway.exception.ClientException;
+import wooteco.subway.service.infra.DijkstraPathStrategy;
 
 import java.util.List;
